QUANTIUM TASK 1 (VIRTUAL INTERNSHIP)
Task 1 :This is a virtual internship provided by Quantium and Forage. It consists of data analysis concerning commerce data, in particular the sales of chips. QUANTIUM TASK 1 where we have to data exploration , find the null values if have then remove them and checking the oultliers are there or not by the boxplot then, remove the outliers.